完成后端登录验证

+++ movie-service/src/main/java/top/yeyuchun/service/Impl/MovieServiceImpl.java
@@ -1,5 +1,6 @@
 package top.yeyuchun.service.Impl;
 
+import cn.hutool.core.util.ArrayUtil;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
@@ -15,6 +16,7 @@ import org.springframework.web.reactive.function.client.WebClient;
 import reactor.netty.http.client.HttpClient;
 import reactor.netty.transport.ProxyProvider;
 import top.yeyuchun.entity.Movie;
+import top.yeyuchun.exception.BusinessException;
 import top.yeyuchun.mapper.MovieMapper;
 import top.yeyuchun.service.MovieService;
 
@@ -313,6 +315,9 @@ public class MovieServiceImpl implements MovieService {
     @Override
     @Transactional
     public void deleteById(Integer id) {
+        if(id == null) {
+            throw new BusinessException("请选择要删除的影视");
+        }
         // 1. 先删除中间表
         movieMapper.deleteMovieGenres(id);
 
@@ -323,6 +328,9 @@ public class MovieServiceImpl implements MovieService {
     @Override
     @Transactional
     public void deleteBatch(Integer[] ids) {
+        if(ArrayUtil.isEmpty(ids)) {
+            throw new BusinessException("请选择要删除的影视");
+        }
         // 1. 先删除中间表的所有关联记录
         movieMapper.deleteMovieGenresBatch(ids);
